adds first approach to palette card refactor
