+++ lib/sources/presentation/layouts/names_list/names_list_grid.dart
@@ -4,8 +4,8 @@ import 'package:colored/sources/app/styling/duration/duration_data.dart';
 import 'package:colored/sources/app/styling/padding/padding_data.dart';
 import 'package:colored/sources/common/extensions/hex_color.dart';
 import 'package:colored/sources/domain/data_models/named_color.dart';
+import 'package:colored/sources/presentation/widgets/cards/single_color_card.dart';
 import 'package:colored/sources/presentation/widgets/containers/background_container.dart';
-import 'package:colored/sources/presentation/widgets/containers/color_card.dart';
 import 'package:flutter/material.dart';
 
 const _kEstimatedItemSize = 200;
@@ -60,7 +60,7 @@ class _NamesListGridState extends State<NamesListGrid> {
               crossAxisCount: _computeCrossAxisCount(constraints),
               childAspectRatio: 1,
             ),
-            itemBuilder: (_, i) => ColorCard(
+            itemBuilder: (_, i) => SingleColorCard(
               backgroundColor: HexColor.fromHex(widget.namedColors[i].hex),
               onPressed: widget.onCardPressed,
               title: widget.namedColors[i].name,
